factor out tooltip content
