+++ packages/studio-base/src/components/PlaybackControls/Scrubber.tsx
@@ -2,10 +2,9 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { Divider, Tooltip, Typography } from "@mui/material";
+import { Tooltip } from "@mui/material";
 import { Instance } from "@popperjs/core";
-import { isEmpty } from "lodash";
-import { Fragment, useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { useLatest } from "react-use";
 import { makeStyles } from "tss-react/mui";
 import { v4 as uuidv4 } from "uuid";
@@ -17,42 +16,18 @@ import {
 } from "@foxglove/studio-base/components/MessagePipeline";
 import Stack from "@foxglove/studio-base/components/Stack";
 import {
-  TimelineInteractionStateStore,
   useClearHoverValue,
   useSetHoverValue,
-  useTimelineInteractionState,
 } from "@foxglove/studio-base/context/TimelineInteractionStateContext";
-import { useAppTimeFormat } from "@foxglove/studio-base/hooks";
 import { PlayerPresence } from "@foxglove/studio-base/players/types";
-import { fonts } from "@foxglove/studio-base/util/sharedStyleConstants";
 
 import { EventsOverlay } from "./EventsOverlay";
 import PlaybackBarHoverTicks from "./PlaybackBarHoverTicks";
+import { PlaybackControlsTooltipContent } from "./PlaybackControlsTooltipContent";
 import { ProgressPlot } from "./ProgressPlot";
 import Slider from "./Slider";
 
 const useStyles = makeStyles()((theme) => ({
-  tooltipDivider: {
-    gridColumn: "span 2",
-    marginBlock: theme.spacing(0.5),
-    opacity: 0.5,
-  },
-  tooltipWrapper: {
-    fontFeatureSettings: `${fonts.SANS_SERIF_FEATURE_SETTINGS}, "zero"`,
-    fontFamily: fonts.SANS_SERIF,
-    whiteSpace: "nowrap",
-    columnGap: theme.spacing(0.5),
-    display: "grid",
-    alignItems: "center",
-    gridTemplateColumns: "auto 1fr",
-    flexDirection: "column",
-  },
-  itemKey: {
-    fontSize: "0.7rem",
-    opacity: 0.7,
-    textAlign: "end",
-    textTransform: "lowercase",
-  },
   marker: {
     backgroundColor: theme.palette.text.primary,
     position: "absolute",
@@ -73,7 +48,6 @@ const useStyles = makeStyles()((theme) => ({
   },
 }));
 
-const selectHoveredEvents = (store: TimelineInteractionStateStore) => store.eventsAtHoverValue;
 const selectStartTime = (ctx: MessagePipelineContext) => ctx.playerState.activeData?.startTime;
 const selectCurrentTime = (ctx: MessagePipelineContext) => ctx.playerState.activeData?.currentTime;
 const selectEndTime = (ctx: MessagePipelineContext) => ctx.playerState.activeData?.endTime;
@@ -85,23 +59,18 @@ type Props = {
   onSeek: (seekTo: Time) => void;
 };
 
-type TooltipItem = { type: "divider" } | { type: "item"; title: string; value: string };
-
 export default function Scrubber(props: Props): JSX.Element {
   const { onSeek } = props;
   const { classes, cx } = useStyles();
 
   const [hoverComponentId] = useState<string>(() => uuidv4());
-  const el = useRef<HTMLDivElement>(ReactNull);
-
-  const { formatTime, timeFormat } = useAppTimeFormat();
+  const hoverElRef = useRef<HTMLDivElement>(ReactNull);
 
   const startTime = useMessagePipeline(selectStartTime);
   const currentTime = useMessagePipeline(selectCurrentTime);
   const endTime = useMessagePipeline(selectEndTime);
   const presence = useMessagePipeline(selectPresence);
   const ranges = useMessagePipeline(selectRanges);
-  const hoveredEvents = useTimelineInteractionState(selectHoveredEvents);
 
   const setHoverValue = useSetHoverValue();
 
@@ -111,8 +80,8 @@ export default function Scrubber(props: Props): JSX.Element {
 
   const latestStartTime = useLatest(startTime);
   const onHoverOver = useCallback(
-    (x: number, value: number) => {
-      if (!latestStartTime.current || el.current == undefined) {
+    (_x: number, value: number) => {
+      if (!latestStartTime.current || hoverElRef.current == undefined) {
         return;
       }
       const stamp = fromSec(value);
@@ -132,68 +101,6 @@ export default function Scrubber(props: Props): JSX.Element {
   const onHoverOut = useCallback(() => {
     clearHoverValue(hoverComponentId);
   }, [clearHoverValue, hoverComponentId]);
-
-  const tooltip = useMemo(() => {
-    if (!latestStartTime.current || el.current == undefined || hoverX == undefined) {
-      return;
-    }
-    const stamp = fromSec(hoverX);
-    const timeFromStart = subtractTimes(stamp, latestStartTime.current);
-
-    const tooltipItems: TooltipItem[] = [];
-
-    if (!isEmpty(hoveredEvents)) {
-      Object.values(hoveredEvents).forEach(({ event }) => {
-        tooltipItems.push({
-          type: "item",
-          title: "Start",
-          value: formatTime(event.startTime),
-        });
-        tooltipItems.push({
-          type: "item",
-          title: "End",
-          value: formatTime(event.endTime),
-        });
-        if (!isEmpty(event.metadata)) {
-          Object.entries(event.metadata).forEach(([metaKey, metaValue]) => {
-            tooltipItems.push({ type: "item", title: metaKey, value: metaValue });
-          });
-        }
-        tooltipItems.push({ type: "divider" });
-      });
-    }
-
-    switch (timeFormat) {
-      case "TOD":
-        tooltipItems.push({ type: "item", title: "Time", value: formatTime(stamp) });
-        break;
-      case "SEC":
-        tooltipItems.push({ type: "item", title: "SEC", value: formatTime(stamp) });
-        break;
-    }
-
-    tooltipItems.push({
-      type: "item",
-      title: "Elapsed",
-      value: `${toSec(timeFromStart).toFixed(9)} sec`,
-    });
-
-    return (
-      <div className={classes.tooltipWrapper}>
-        {tooltipItems.map((item, idx) => {
-          if (item.type === "divider") {
-            return <Divider key={`divider_${idx}`} className={classes.tooltipDivider} />;
-          }
-          return (
-            <Fragment key={`${item.title}_${idx}`}>
-              <Typography className={classes.itemKey}>{item.title}</Typography>
-              <Typography variant="subtitle2">{item.value}</Typography>
-            </Fragment>
-          );
-        })}
-      </div>
-    );
-  }, [classes, formatTime, hoveredEvents, latestStartTime, timeFormat, hoverX]);
 
   // Clean up the hover value when we are unmounted -- important for storybook.
   useEffect(() => onHoverOut, [onHoverOut]);
@@ -217,13 +124,9 @@ export default function Scrubber(props: Props): JSX.Element {
 
   const popperRef = React.useRef<Instance>(ReactNull);
 
-  const areaRef = useRef<HTMLDivElement>(ReactNull);
-  const positionRef = React.useRef<{ x: number; y: number }>({
-    x: 0,
-    y: 0,
-  });
+  const positionRef = React.useRef({ x: 0, y: 0 });
 
-  const handleMouseMove = (event: React.MouseEvent) => {
+  const handlePointerMove = (event: React.PointerEvent) => {
     positionRef.current = { x: event.clientX, y: event.clientY };
 
     if (popperRef.current != undefined) {
@@ -233,7 +136,7 @@ export default function Scrubber(props: Props): JSX.Element {
 
   return (
     <Tooltip
-      title={<>{tooltip}</>}
+      title={hoverX != undefined ? <PlaybackControlsTooltipContent hoverXPosition={hoverX} /> : ""}
       placement="top"
       PopperProps={{
         popperRef,
@@ -241,7 +144,7 @@ export default function Scrubber(props: Props): JSX.Element {
           getBoundingClientRect: () => {
             return new DOMRect(
               positionRef.current.x,
-              areaRef.current!.getBoundingClientRect().y,
+              8 + (hoverElRef.current?.getBoundingClientRect().y ?? 0),
               0,
               0,
             );
@@ -250,18 +153,19 @@ export default function Scrubber(props: Props): JSX.Element {
       }}
     >
       <Stack
+        ref={hoverElRef}
         direction="row"
         flexGrow={1}
-        onMouseMove={handleMouseMove}
+        onPointerMove={handlePointerMove}
         alignItems="center"
         position="relative"
         style={{ height: 32 }}
       >
         <div className={cx(classes.track, { [classes.trackDisabled]: !startTime })} />
-        <Stack ref={areaRef} position="absolute" flex="auto" fullWidth style={{ height: 6 }}>
+        <Stack position="absolute" flex="auto" fullWidth style={{ height: 6 }}>
           <ProgressPlot loading={loading} availableRanges={ranges} />
         </Stack>
-        <Stack ref={el} fullHeight fullWidth position="absolute" flex={1}>
+        <Stack fullHeight fullWidth position="absolute" flex={1}>
           <Slider
             min={min ?? 0}
             max={max ?? 100}
